new changes

+++ src/validation/schemas/cartSchemas.js
@@ -3,5 +3,5 @@ const { body } = require('express-validator');
 
 exports.addToCart = [
   body('menuItemId').isMongoId(),
-  body('quantity').optional().isInt({ min: 1, max: 50 })
+  body('quantity').optional().isInt({ min: 1, max: 50 }).toInt()
 ];
